refactor: 쿠키 스냅샷 불러오기 로직을 추상화
