+++ src/shared/lib/auth/crypto.ts
